Move BMP390 values to floats

+++ Core/Inc/bmp390_runner.h
@@ -26,8 +26,8 @@
 
 /** Public variables. *********************************************************/
 
-extern double bmp390_temperature;
-extern double bmp390_pressure;
+extern float bmp390_temperature;
+extern float bmp390_pressure;
 
 /** Public functions. *********************************************************/
